+++ stream/auth/stripe/functions/createStripeClient.js
@@ -1,0 +1,7 @@
+import Stripe from 'stripe';
+
+export default function createStripeClient(STRIPE_SECRET_KEY) {
+	return Stripe(STRIPE_SECRET_KEY, {
+		httpClient: Stripe.createFetchHttpClient(),
+	});
+}
