+++ stream/auth/stripe/src/canceled.html
@@ -3,11 +3,7 @@
 	<head>
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-
-		<title>Stripe Checkout Sample</title>
-
-		<link rel="icon" href="favicon.ico" type="image/x-icon" />
-		<link rel="stylesheet" href="css/normalize.css" />
+		<title>Cloudflare Stream + Stripe Checkout Example — Payment Canceled</title>
 		<link rel="stylesheet" href="css/global.css" />
 	</head>
 
@@ -20,14 +16,6 @@
 				<div class="sr-payment-summary completed-view">
 					<h1>Your payment was canceled</h1>
 					<button onclick="window.location.href = '/';">Restart demo</button>
-				</div>
-			</div>
-			<div class="sr-content">
-				<div class="pasha-image-stack">
-					<img src="https://picsum.photos/280/320?random=1" width="140" height="160" />
-					<img src="https://picsum.photos/280/320?random=2" width="140" height="160" />
-					<img src="https://picsum.photos/280/320?random=3" width="140" height="160" />
-					<img src="https://picsum.photos/280/320?random=4" width="140" height="160" />
 				</div>
 			</div>
 		</div>
