+++ stream/auth/stripe/src/css/watch.css
@@ -1,0 +1,53 @@
+html {
+    background: #222;
+}
+
+html,
+body {
+    margin: 0px;
+}
+
+body {
+    min-height: 100vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+section {
+    --video-aspect-ratio: 16 / 9;
+    --padding: 2em;
+
+    box-sizing: border-box;
+    width: calc((100vh - (2 * var(--padding))) * var(--video-aspect-ratio));
+    max-width: 100%;
+    padding: var(--padding);
+}
+
+stream {
+    display: block;
+    width: 1080px;
+    max-width: 100%;
+    margin: 0 auto;
+    box-shadow: 0 0 30px #000;
+    box-shadow: 0 0 5vmin #000;
+}
+
+@media (max-width: 768px) {
+    html {
+        background: #000;
+    }
+
+    section {
+        --padding: 0px;
+    }
+
+    stream {
+        box-shadow: none;
+    }
+}
+
+stream {
+    position: relative;
+    display: block;
+}
