+++ stream/auth/stripe/src/watch.html
@@ -3,68 +3,10 @@
 	<head>
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-
 		<title>Cloudflare Stream + Stripe — Authenticated Viewing Page</title>
-
+		<link rel="stylesheet" href="css/watch.css" />
 		<script src="./watch.js" defer></script>
-		<style>
-			html {
-				background: #222;
-			}
-
-			html,
-			body {
-				margin: 0px;
-			}
-
-			body {
-				min-height: 100vh;
-				display: flex;
-				align-items: center;
-				justify-content: center;
-			}
-
-			section {
-				--video-aspect-ratio: 16 / 9;
-				--padding: 2em;
-
-				box-sizing: border-box;
-				width: calc((100vh - (2 * var(--padding))) * var(--video-aspect-ratio));
-				max-width: 100%;
-				padding: var(--padding);
-			}
-
-			stream {
-				display: block;
-				width: 1080px;
-				max-width: 100%;
-				margin: 0 auto;
-				box-shadow: 0 0 30px #000;
-				box-shadow: 0 0 5vmin #000;
-			}
-
-			@media (max-width: 768px) {
-				html {
-					background: #000;
-				}
-
-				section {
-					--padding: 0px;
-				}
-
-				stream {
-					box-shadow: none;
-				}
-			}
-		</style>
-		<style>
-			stream {
-				position: relative;
-				display: block;
-			}
-		</style>
 	</head>
-
 	<body>
 		<section>
 			<div
